+++ packages/typespec-rust/test/spector/azure/resource-manager/method-subscription-id/tests/method_subscription_id_mixed_subscription_placement_resource_group_resource_operations_client_test.rs
@@ -21,7 +21,7 @@ async fn resource_group_resource_get() {
         .await
         .unwrap();
 
-    let resource: ResourceGroupResource = resp.into_body().await.unwrap();
+    let resource: ResourceGroupResource = resp.into_body().unwrap();
     let expected_resource = get_valid_resource_group_resource();
 
     assert_eq!(expected_resource.id, resource.id);
@@ -55,7 +55,7 @@ async fn resource_group_resource_put() {
         .await
         .unwrap();
 
-    let created_resource: ResourceGroupResource = resp.into_body().await.unwrap();
+    let created_resource: ResourceGroupResource = resp.into_body().unwrap();
 
     assert_eq!(resource.location, created_resource.location);