+++ packages/typespec-rust/test/spector/encode/duration/tests/duration_property_client_test.rs
@@ -21,7 +21,7 @@ async fn default() {
         .default(body.try_into().unwrap(), None)
         .await
         .unwrap();
-    let result: DefaultDurationProperty = resp.into_body().await.unwrap();
+    let result: DefaultDurationProperty = resp.into_body().unwrap();
     assert_eq!(result.value, Some("P40D".to_string()));
 }
 
@@ -36,7 +36,7 @@ async fn float64_seconds() {
         .float64_seconds(body.try_into().unwrap(), None)
         .await
         .unwrap();
-    let result: Float64SecondsDurationProperty = resp.into_body().await.unwrap();
+    let result: Float64SecondsDurationProperty = resp.into_body().unwrap();
     assert_eq!(result.value, Some(35.625));
 }
 
@@ -51,7 +51,7 @@ async fn float_seconds() {
         .float_seconds(body.try_into().unwrap(), None)
         .await
         .unwrap();
-    let result: FloatSecondsDurationProperty = resp.into_body().await.unwrap();
+    let result: FloatSecondsDurationProperty = resp.into_body().unwrap();
     assert_eq!(result.value, Some(35.625));
 }
 
@@ -66,7 +66,7 @@ async fn float_seconds_array() {
         .float_seconds_array(body.try_into().unwrap(), None)
         .await
         .unwrap();
-    let result: FloatSecondsDurationArrayProperty = resp.into_body().await.unwrap();
+    let result: FloatSecondsDurationArrayProperty = resp.into_body().unwrap();
     assert_eq!(result.value, Some(vec![35.625, 46.75]));
 }
 
@@ -79,7 +79,7 @@ async fn int32_seconds() {
         .int32_seconds(body.try_into().unwrap(), None)
         .await
         .unwrap();
-    let result: Int32SecondsDurationProperty = resp.into_body().await.unwrap();
+    let result: Int32SecondsDurationProperty = resp.into_body().unwrap();
     assert_eq!(result.value, Some(36));
 }
 
@@ -94,6 +94,6 @@ async fn iso8601() {
         .iso8601(body.try_into().unwrap(), None)
         .await
         .unwrap();
-    let result: ISO8601DurationProperty = resp.into_body().await.unwrap();
+    let result: ISO8601DurationProperty = resp.into_body().unwrap();
     assert_eq!(result.value, Some("P40D".to_string()));
 }