+++ packages/typespec-rust/test/spector/payload/json-merge-patch/tests/json_merge_patch_client_test.rs
@@ -34,7 +34,7 @@ async fn create_resource() {
         .await
         .unwrap();
 
-    let value: Resource = resp.into_body().await.unwrap();
+    let value: Resource = resp.into_body().unwrap();
     assert_eq!(value.name, Some(String::from("Madge")));
     assert_eq!(value.description, Some(String::from("desc")));