+++ packages/typespec-rust/test/spector/type/property/optionality/tests/optional_duration_client.rs
@@ -13,7 +13,6 @@ async fn get_all() {
         .await
         .unwrap()
         .into_body()
-        .await
         .unwrap();
     // According to mockapi.ts, the all endpoint returns { property: "P123DT22H14M12.011S" }
     assert_eq!(resp.property, Some("P123DT22H14M12.011S".to_string()));
@@ -28,7 +27,6 @@ async fn get_default() {
         .await
         .unwrap()
         .into_body()
-        .await
         .unwrap(); // According to mockapi.ts, the default endpoint returns {}
     assert!(resp.property.is_none());
 }