+++ packages/typespec-rust/test/spector/versioning/madeOptional/tests/made_optional_client.rs
@@ -16,7 +16,6 @@ async fn test() {
         .await
         .unwrap()
         .into_body()
-        .await
         .unwrap();
     assert_eq!(resp.prop, Some("foo".to_string()));
 }